requirements.txt
streamlit==1.38.0
geopy==2.4.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
folium==0.17.0
streamlit-folium==0.20.0
SpeechRecognition==3.10.4
PyAudio==0.2.14
Pillow==10.4.0
